\name{Kitis}
\alias{Kitis}
\docType{data}
\title{
Glow curves provided by George Kitis
}

\description{
A total of 22 glow curves provided by George Kitis.
}

\usage{data(Kitis)}

\format{
A list that contains 22 thermoluminescence glow curves.
}

\details{
This object contains 22 thermoluminescence glow curves (\code{x001} to \code{x022}) provided by George Kitis.
}

\examples{
 # Load package "tgcd".
  require(tgcd)

  data(Kitis)
  names(Kitis)
}

\keyword{package tgcd}
\keyword{Thermoluminescence}
\keyword{Glow curve simulation}
\keyword{Glow curve deconvolution}
