\name{tgcd-package}
\alias{tgcd-package}
\docType{package}
\title{tgcd: An R package for analyzing thermoluminescence glow curves}
\description{
Package for thermoluminescence glow curve deconvolution (tgcd) and glow peak simulation.
}
\details{
\tabular{ll}{
Package: \tab tgcd\cr
Type: \tab Package\cr
Version: \tab 2.6\cr
Date: \tab 2023-08-29\cr
License: \tab  GPL-2 | GPL-3 \cr
}
}
\author{
\tabular{ll}{
Jun Peng \tab Hunan University of Science and Technology, Xiangtan, China
}

\bold{Package maintainer}\cr
Jun Peng
\email{pengjun10@mails.ucas.ac.cn}

\bold{Related package projects}\cr
R program KMS
\url{https://github.com/pengjunUCAS/KMS}\cr
R package numOSL
\url{https://CRAN.R-project.org/package=numOSL}
}
\references{
Peng J, Dong ZB, Han FQ, 2016. tgcd: An R package for analyzing thermoluminescence glow curves. SoftwareX, 5: 112-120.

Peng J, Kitis G, Sadek AM, Karsu Asal EC, Li, ZG, 2021. Thermoluminescence glow-curve deconvolution using analytical
expressions: A unified presentation. Applied Radiation and Isotopes, 168: 109440.
}
\keyword{Thermoluminescence}
\keyword{Glow curve}
\keyword{Glow peak}
\keyword{Deconvolution}
\keyword{Simulation}
